+++ Common/src/main/java/fuzs/easymagic/mixin/accessor/PlayerAccessor.java
@@ -6,6 +6,7 @@ import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(Player.class)
 public interface PlayerAccessor {
+
     @Accessor
     void setEnchantmentSeed(int enchantmentSeed);
 }